split reports data logic by skin type
